Implement a simple signal handler for SIGINT
Most of the time is spent in the play_wave function, thus it has to be interrupted. A flag together with a call to audio_drain() is used to cancel playback and EINTR is returned when exiting due to SIGINT.

mimic_process_output() has been updated to return dur via an argument and normal return is now an error value.