ALSA and PortAudio detection using pkg-config
Let pkg-config define the right CFLAGS and LDFLAGS for PortAudio and ALSA
detection. Further cleaning of configure script will come in future commits.

pkg-config makes sure all the variables needed for finding headers and linking
are provided, no matter the OS.

If pkg-config is not available or if the user wants to, it is possible to use
ALSA_CFLAGS ALSA_LIBS PORTAUDIO_CFLAGS and PORTAUDIO_LIBS to specify the
respective required variables.